Add SOS emoji
SOS is the International Morse code distress signal (▄▄▄▄▄▄▄▄▄▄▄▄▄▄▄▄▄▄▄▄▄▄▄▄▄▄). This distress signal was first adopted by the German government radio regulations effective April 1, 1905, and became the worldwide standard under the second International Radiotelegraphic Convention, which was signed on November 3, 1906, and became effective on July 1, 1908. SOS remained the maritime radio distress signal until 1999, when it was replaced by the Global Maritime Distress and Safety System. SOS is still recognized as a visual distress signal.

The SOS distress signal is a continuous spaceless sequence of three dots, three dashes, and three dots. In International Morse Code, three dots form the letter S, and three dashes make the letter O, so "SOS" became an easy way to remember the order of the dots and dashes. In modern terminology, SOS is a Morse "procedural signal" or "prosign", and the formal way to write it is with a bar above the letters: SOS.

In popular usage, SOS became associated with such phrases as "Save our Souls". SOS is only one of several ways that the combination could have been written; VTB, for example, would produce exactly the same sound, but SOS was chosen to describe this combination. SOS is the only nine-element signal in Morse code, making it more easily recognizable, as no other symbol uses more than eight elements.