Add Latin cross
A latin cross, otherwise known as a Christian cross. This cross symbol is often used as a symbol of Christianity. Similar in appearance to the Orthodox cross.

Latin Cross was approved as part of Unicode 1.1 in 1993 and added to Emoji 1.0 in 2015.

This is a list of Christian cross variants. The Christian cross, with or without a figure of Christ included, is the main religious symbol of Christianity. A cross with figure of Christ affixed to it is termed a crucifix and the figure is often referred to as the corpus (Latin for "body").

The term Greek cross designates a cross with arms of equal length, as in a plus sign, while the term Latin cross designates a cross with an elongated descending arm. Numerous other variants have been developed during the medieval period.

Christian crosses are used widely in churches, on top of church buildings, on bibles, in heraldry, in personal jewelry, on hilltops, and elsewhere as an attestation or other symbol of Christianity. Crosses are a prominent feature of Christian cemeteries, either carved on gravestones or as sculpted stelae. Because of this, planting small crosses is sometimes used in countries of Christian culture to mark the site of fatal accidents, ors, such as the Zugspitze or Mount Royal, so as to be visible over the entire surrounding area. Roman Catholic, Anglican and Lutheran depictions of the cross are often crucifixes, in order to emphasize that it is Jesus that is important, rather than the cross in isolation. Large crucifixes are a prominent feature of some Lutheran churches, as illustrated in the article Rood. However, some other Protestant traditions depict the cross without the corpus, interpreting this form as an indication of belief in the resurrection rather than as representing the interval between the death and the resurrection of Jesus.

Several Christian cross variants are available in computer-displayed text. The Latin cross symbol ("✝") is included in the unicode character set as "271D". For others, see Religious and political symbols in Unicode.